✨ feat(app): add RestaurantAuthorizationService

+++ Restaurants.Application/Dishes/Commands/DeleteDish/DeleteDishesRestaurantCommandHandler.cs
@@ -1,14 +1,16 @@
-using AutoMapper;
 using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Dishes.Commands.DeleteDish;
 public class DeleteDishesRestaurantCommandHandler(
     ILogger<DeleteDishesRestaurantCommandHandler> logger,
     IRestaurantsRepository restaurantsRepository,
+    IRestaurantAuthorizationService restaurantAuthorizationService,
     IDishesRepository dishesRepository
     ) : IRequestHandler<DeleteDishesRestaurantCommand>
 {
@@ -18,6 +20,11 @@ public class DeleteDishesRestaurantCommandHandler(
 
         var restaurant = await restaurantsRepository.GetByIdAsync(request.RestaurantId);
         if (restaurant is null) throw new NotFoundException(nameof(Restaurant), request.RestaurantId.ToString());
+
+        if (!restaurantAuthorizationService.Authorize(restaurant, ResourceOperation.Delete))
+        {
+            throw new ForbiddenException();
+        }
 
         await dishesRepository.Delete(restaurant.Dishes);
     }
